Update many_bins and move it to regression
